:seedling: Move remaining rest functions to dedicated files (no logic changes)
Move all remaining rest function groups out of rest.ts into individual files
under client/src/app/api/rest/. Functions are copied verbatim — signatures,
return types, and implementations are unchanged from rest.ts.

Part of #2630

New files:
- rest/assessments.ts
- rest/facts.ts
- rest/imports.ts       (includes APP_IMPORTS_SUMMARY_UPLOAD export)
- rest/proxies.ts
- rest/questionnaires.ts (includes QUESTIONNAIRES export)
- rest/reviews.ts
- rest/settings.ts
- rest/stakeholder-groups.ts
- rest/stakeholders.ts
- rest/tags.ts          (includes tag categories)
- rest/targets.ts
- rest/task-groups.ts
- rest/tickets.ts
- rest/trackers.ts

rest.ts now only contains: the hub() helper, HEADERS, FILES, and barrel
re-exports for all rest/* modules.
