+++ client/src/app/api/rest/trackers.ts
@@ -1,0 +1,35 @@
+import axios from "axios";
+
+import { Tracker, TrackerProject, TrackerProjectIssuetype } from "../models";
+import { hub } from "../rest";
+
+const TRACKERS = hub`/trackers`;
+const TRACKER_PROJECT_ISSUETYPES = "issuetypes"; // TODO: ????
+const TRACKER_PROJECTS = "projects"; // TODO: ????
+
+export const getTrackers = (): Promise<Tracker[]> =>
+  axios.get(TRACKERS).then((response) => response.data);
+
+export const createTracker = (obj: Tracker): Promise<Tracker> =>
+  axios.post(TRACKERS, obj).then((res) => res.data);
+
+export const updateTracker = (obj: Tracker): Promise<Tracker> =>
+  axios.put(`${TRACKERS}/${obj.id}`, obj);
+
+export const deleteTracker = (id: number): Promise<Tracker> =>
+  axios.delete(`${TRACKERS}/${id}`);
+
+export const getTrackerProjects = (id: number): Promise<TrackerProject[]> =>
+  axios
+    .get(`${TRACKERS}/${id}/${TRACKER_PROJECTS}`)
+    .then((response) => response.data);
+
+export const getTrackerProjectIssuetypes = (
+  trackerId: number,
+  projectId: string
+): Promise<TrackerProjectIssuetype[]> =>
+  axios
+    .get(
+      `${TRACKERS}/${trackerId}/${TRACKER_PROJECTS}/${projectId}/${TRACKER_PROJECT_ISSUETYPES}`
+    )
+    .then((response) => response.data);
